[PSL-314] file size round up for action fee calculation

+++ walletnode/services/cascaderegister/service.go
@@ -105,7 +105,7 @@ func (service *CascadeRegistrationService) ValidateDetailsAndCalculateFee(ctx co
 		return 0.0, err
 	}
 
-	fileDataInMb := int64(len(fileData)) / (1024 * 1024)
+	fileDataInMb := float64(len(fileData)) / (1024 * 1024)
 
 	// Validate image signature
 	ok, err := service.pastelHandler.VerifySignature(ctx,
